+++ packages/terra-form-input/src/terra-dev-site/test/form-input/DisabledFormInput.test.jsx
@@ -1,6 +1,14 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import Input from '../../../Input';
+import styles from './common/Input.test.module.scss';
 
-const DisabledInput = () => <Input disabled id="form-input-disabled" ariaLabel="label" />;
+const cx = classNames.bind(styles);
+
+const DisabledInput = () => (
+  <div className={cx('content-wrapper')}>
+    <Input disabled id="form-input-disabled" ariaLabel="label" />
+  </div>
+);
 
 export default DisabledInput;