+++ packages/terra-form-select/src/ComboboxField.jsx
@@ -39,11 +39,15 @@ const propTypes = {
    */
   hideRequired: PropTypes.bool,
   /**
+   * Whether the field displays as Incomplete. Use when no value has been provided. _(usage note: `required` must also be set)_.
+   */
+  isIncomplete: PropTypes.bool,
+  /**
    * Whether the field is displayed inline. Displays block by default.
    */
   isInline: PropTypes.bool,
   /**
-   * Whether the field is invalid.
+   * Whether the field displays as Invalid. Use when value does not meet validation pattern.
    */
   isInvalid: PropTypes.bool,
   /**
@@ -105,6 +109,7 @@ const defaultProps = {
   error: undefined,
   help: undefined,
   hideRequired: false,
+  isIncomplete: false,
   isInline: false,
   isInvalid: false,
   isTouchAccessible: false,
@@ -127,6 +132,7 @@ const ComboboxField = ({
   error,
   help,
   hideRequired,
+  isIncomplete,
   isInline,
   isInvalid,
   isLabelHidden,
@@ -180,6 +186,7 @@ const ComboboxField = ({
         aria-describedby={ariaDescriptionIds}
         disabled={selectAttrs.disabled || disabled}
         id={selectId}
+        isIncomplete={isIncomplete}
         isInvalid={isInvalid}
         isTouchAccessible={isTouchAccessible}
         defaultValue={defaultValue}