+++ packages/terra-form-select/src/terra-dev-site/doc/form-select/MultiSelect.07.doc.jsx
@@ -20,6 +20,8 @@ import IsTouchAccessibleExample from '../example/multiple/IsTouchAccessible';
 import IsTouchAccessibleExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/multiple/IsTouchAccessible';
 import InvalidExample from '../example/multiple/Invalid';
 import InvalidExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/multiple/Invalid';
+import IncompleteExample from '../example/multiple/Incomplete';
+import IncompleteExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/multiple/Incomplete';
 import MaxHeightExample from '../example/multiple/MaxHeight';
 import MaxHeightExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/multiple/MaxHeight';
 import OptGroupExample from '../example/multiple/OptGroup';
@@ -67,6 +69,11 @@ const Index = () => (
       description: 'Applies theme-specific styling for invalid. Will also display error text when used with a Field.',
       example: <InvalidExample />,
       source: InvalidExampleSrc,
+    }, {
+      title: 'Incomplete',
+      description: 'Applies theme-specific styling for incomplete. ***Note: Only use incomplete if given specific guidance, reserved for specific applications when no value has been provided. Not for general use.***',
+      example: <IncompleteExample />,
+      source: IncompleteExampleSrc,
     }, {
       title: 'Ability to add Option Groups',
       description: 'An OptGroup provides a way to group options together within the listbox.',