+++ packages/terra-form-select/tests/jest/Frame.test.jsx
@@ -108,6 +108,31 @@ describe('Frame', () => {
     expect(wrapper).toMatchSnapshot();
   });
 
+  it('should render an incomplete default variant', () => {
+    const wrapper = shallowWithIntl(<SingleSelectFrame isIncomplete />);
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('should render an incomplete combobox variant', () => {
+    const wrapper = shallowWithIntl(<ComboboxFrame isIncomplete />);
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('should render an incomplete multiple variant', () => {
+    const wrapper = shallowWithIntl(<MultipleFrame isIncomplete />);
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('should render an incomplete search variant', () => {
+    const wrapper = shallowWithIntl(<SearchFrame isIncomplete />);
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('should render an incomplete tag variant', () => {
+    const wrapper = shallowWithIntl(<TagFrame isIncomplete />);
+    expect(wrapper).toMatchSnapshot();
+  });
+
   it('should render a custom dropdown', () => {
     const wrapper = shallowWithIntl(<TagFrame dropdown={() => <div>Custom</div>} />);
     expect(wrapper).toMatchSnapshot();