+++ packages/terra-form-textarea/src/terra-dev-site/test/form-textarea/TextareaField.test.jsx
@@ -1,5 +1,9 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import TextareaField from '../../../TextareaField';
+import styles from './common/Textarea.test.module.scss';
+
+const cx = classNames.bind(styles);
 
 class TextareaFieldExample extends React.Component {
   constructor() {
@@ -7,9 +11,12 @@ class TextareaFieldExample extends React.Component {
 
     this.state = {
       isInvalid: false,
+      isIncomplete: false,
+      required: false,
     };
 
     this.toggleInvalid = this.toggleInvalid.bind(this);
+    this.toggleIncomplete = this.toggleIncomplete.bind(this);
   }
 
   toggleInvalid() {
@@ -18,10 +25,18 @@ class TextareaFieldExample extends React.Component {
     }));
   }
 
+  toggleIncomplete() {
+    this.setState(prevState => ({
+      isIncomplete: !prevState.isIncomplete,
+      required: !prevState.required,
+    }));
+  }
+
   render() {
     return (
-      <div>
+      <div className={cx('content-wrapper')}>
         <button type="button" id="validity-toggle" onClick={this.toggleInvalid}>Toggle Validity</button>
+        <button type="button" id="incomplete-toggle" onClick={this.toggleIncomplete}>Toggle Incomplete</button>
         <TextareaField
           defaultValue="Value"
           inputId="test-input"
@@ -32,6 +47,8 @@ class TextareaFieldExample extends React.Component {
             name: 'test',
           }}
           isInvalid={this.state.isInvalid}
+          isIncomplete={this.state.isIncomplete}
+          required={this.state.required}
           label="Label Text"
           labelAttrs={{
             className: 'label',